add error handlers on clientSide and serverSide

+++ Library.Domain/Concrete/MongoDbConcrete/AuthorMongoDbRepository.cs
@@ -2,6 +2,7 @@
 using Library.Domain.Entities;
 using MongoDB.Bson;
 using MongoDB.Driver;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 namespace Library.Domain.Concrete
@@ -35,40 +36,62 @@ namespace Library.Domain.Concrete
  
         public async Task<int> CreateAuthor(Author author)
         {
-             int DbResult = 0;
-             if (dataReqiered.IsDataRequered(author))
-             {
-                AuthorMongoDb newAuthor = new AuthorMongoDb{ Id = author.Id, Name = author.Name, Surname = author.Surname };
-                await db.Authors.InsertOneAsync(newAuthor);
-                DbResult = 1;
-             }
-             return DbResult;
+            int DbResult = 0;
+            if (dataReqiered.IsDataNoEmpty(author))
+            {
+                AuthorMongoDb newAuthor = new AuthorMongoDb { Id = author.Id, Name = author.Name, Surname = author.Surname };
+                try
+                {
+                    await db.Authors.InsertOneAsync(newAuthor);
+                    DbResult = 1;
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                    return DbResult;
+                }
+            }
+            return DbResult;
         }
 
         public async Task<int> UpdateAuthor(string authorId, Author author)
         {
-              int DbResult = 0;
-              AuthorMongoDb oldAuthorData = await db.Authors.Find(new BsonDocument("_id", new ObjectId(authorId))).FirstOrDefaultAsync();
-              if (oldAuthorData != null && dataReqiered.IsDataRequered(author))
-              {
-                 AuthorMongoDb newAuthorData = new AuthorMongoDb { Id = author.Id, Name = author.Name, Surname = author.Surname };
-                 await db.Authors.ReplaceOneAsync(new BsonDocument("_id", new ObjectId(authorId)), newAuthorData);
-                 DbResult = 1;
-              }
-              return DbResult;
+            int DbResult = 0;
+            if (!String.IsNullOrEmpty(authorId) && dataReqiered.IsDataNoEmpty(author))
+            {
+                AuthorMongoDb oldAuthorData = await db.Authors.Find(new BsonDocument("_id", new ObjectId(authorId))).FirstOrDefaultAsync();
+                if (oldAuthorData != null)
+                {
+                    AuthorMongoDb newAuthorData = new AuthorMongoDb { Id = author.Id, Name = author.Name, Surname = author.Surname };
+                    try
+                    {
+                        await db.Authors.ReplaceOneAsync(new BsonDocument("_id", new ObjectId(authorId)), newAuthorData);
+                        DbResult = 1;
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine(ex.Message);
+                        return DbResult;
+                    }
+                }
+            }
+            return DbResult;
         }
 
         public async Task<int> DeleteAuthor(string authorId)
         {
             int DbResult = 0;
-            List<AuthorMongoDb> CollectionResult = await db.Authors.Find(new BsonDocument("_id", new ObjectId(authorId))).ToListAsync();
-            mongoDbDataConvert.InitData(CollectionResult);
-            IEnumerable<Author> deletingAuthor = mongoDbDataConvert.GetIEnumerubleDbResult();
-
-            if (deletingAuthor != null)
+            if (!String.IsNullOrEmpty(authorId))
             {
-                await db.Authors.DeleteOneAsync(new BsonDocument("_id", new ObjectId(authorId)));
-                DbResult = 1;
+                List<AuthorMongoDb> CollectionResult = await db.Authors.Find(new BsonDocument("_id", new ObjectId(authorId))).ToListAsync();
+                mongoDbDataConvert.InitData(CollectionResult);
+                IEnumerable<Author> deletingAuthor = mongoDbDataConvert.GetIEnumerubleDbResult();
+
+                if (deletingAuthor != null)
+                {
+                    await db.Authors.DeleteOneAsync(new BsonDocument("_id", new ObjectId(authorId)));
+                    DbResult = 1;
+                }
             }
             return DbResult;
         }
