+++ Library.Domain/Concrete/MongoDbConcrete/BookMongoDbRepository.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using Library.Domain.Entities;
 using MongoDB.Driver;
 using MongoDB.Bson;
+using System;
 
 namespace Library.Domain.Concrete
 {
@@ -37,11 +38,19 @@ namespace Library.Domain.Concrete
         public async Task<int> CreateBook(Book book)
         {
             int DbResult = 0;
-            if (dataReqiered.IsDataRequered(book))
+            if (dataReqiered.IsDataNoEmpty(book))
             {
                 BookMongoDb newBook = new BookMongoDb { Id = book.Id, Year = book.Year, Name = book.Name, Description = book.Description, AuthorId = book.AuthorId };
-                await db.Books.InsertOneAsync(newBook);
-                DbResult = 1;
+                try
+                {
+                    await db.Books.InsertOneAsync(newBook);
+                    DbResult = 1;
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                    return DbResult;
+                }
             }
             return DbResult;
         }
@@ -49,12 +58,23 @@ namespace Library.Domain.Concrete
         public async Task<int> UpdateBook(string bookId, Book book)
         {
             int DbResult = 0;
-            List<BookMongoDb> oldBookData = await db.Books.Find(new BsonDocument("_id", new ObjectId(bookId))).ToListAsync();
-            if (oldBookData != null && dataReqiered.IsDataRequered(book))
+            if (!String.IsNullOrEmpty(bookId) && dataReqiered.IsDataNoEmpty(book))
             {
-                BookMongoDb newBookData = new BookMongoDb { Id = book.Id, Year = book.Year, Name = book.Name, Description = book.Description, AuthorId = book.AuthorId };
-                await db.Books.ReplaceOneAsync(new BsonDocument("_id", new ObjectId(bookId)), newBookData);
-                DbResult = 1;
+                List<BookMongoDb> oldBookData = await db.Books.Find(new BsonDocument("_id", new ObjectId(bookId))).ToListAsync();
+                if (oldBookData != null)
+                {
+                    BookMongoDb newBookData = new BookMongoDb { Id = book.Id, Year = book.Year, Name = book.Name, Description = book.Description, AuthorId = book.AuthorId };
+                    try
+                    {
+                        await db.Books.ReplaceOneAsync(new BsonDocument("_id", new ObjectId(bookId)), newBookData);
+                        DbResult = 1;
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine(ex.Message);
+                        return DbResult;
+                    }
+                }
             }
             return DbResult;
         }
@@ -62,22 +82,28 @@ namespace Library.Domain.Concrete
         public async Task<int> DeleteBook(string bookId)
         {
              int DbResult = 0;
-             List<BookMongoDb> deletingBook = await db.Books.Find(new BsonDocument("_id", new ObjectId(bookId))).ToListAsync();
-             if (deletingBook != null)
+             if (!String.IsNullOrEmpty(bookId))
              {
-                await db.Books.DeleteOneAsync(new BsonDocument("_id", new ObjectId(bookId)));
-                DbResult = 1;
+                List<BookMongoDb> deletingBook = await db.Books.Find(new BsonDocument("_id", new ObjectId(bookId))).ToListAsync();
+                if (deletingBook != null)
+                {
+                    await db.Books.DeleteOneAsync(new BsonDocument("_id", new ObjectId(bookId)));
+                    DbResult = 1;
+                }
              }
              return DbResult;
         }
 
         public async Task<IEnumerable<Book>> GetBookByAuthorId(string authorId)
         {
-             List<BookMongoDb> BooksByAuthor = await db.Books.Find(new BsonDocument("AuthorId", authorId)).ToListAsync();
-             if(BooksByAuthor != null)
+             if (!String.IsNullOrEmpty(authorId))
              {
-                 mongoDbDataConvert.InitData(BooksByAuthor);
-                 result = mongoDbDataConvert.GetIEnumerubleDbResult();
+                List<BookMongoDb> BooksByAuthor = await db.Books.Find(new BsonDocument("AuthorId", authorId)).ToListAsync();
+                if (BooksByAuthor != null)
+                {
+                    mongoDbDataConvert.InitData(BooksByAuthor);
+                    result = mongoDbDataConvert.GetIEnumerubleDbResult();
+                }
              }
              return result;
         }
